Split new code into a separate file
